revert restructure for another pr
