+++ atuin/src/command/client/search.rs
@@ -11,7 +11,14 @@ use atuin_client::{
 
 use super::history::ListMode;
 
-pub use crate::utils::duration::format_duration_into;
+mod cursor;
+mod duration;
+mod engines;
+mod history_list;
+mod inspector;
+mod interactive;
+
+pub use duration::format_duration_into;
 
 #[allow(clippy::struct_excessive_bools, clippy::struct_field_names)]
 #[derive(Parser, Debug)]
@@ -136,7 +143,7 @@ impl Cmd {
         settings.shell_up_key_binding = self.shell_up_key_binding;
 
         if self.interactive {
-            let item = crate::interactive::search::history(&self.query, settings, db).await?;
+            let item = interactive::history(&self.query, settings, db).await?;
             eprintln!("{item}");
         } else {
             let list_mode = ListMode::from_flags(self.human, self.cmd_only);
