+++ atuin/src/command/client/search/cursor.rs
