+++ atuin/src/command/client/search/interactive.rs
@@ -23,14 +23,14 @@ use atuin_client::{
     settings::{ExitMode, FilterMode, SearchMode, Settings},
 };
 
-use crate::{
+use super::{
+    cursor::Cursor,
     engines::{SearchEngine, SearchState},
-    interactive::{
-        cursor::Cursor,
-        history_list::{HistoryList, ListState, PREFIX_LENGTH},
-    },
-    VERSION,
+    history_list::{HistoryList, ListState, PREFIX_LENGTH},
 };
+
+use crate::{command::client::search::engines, VERSION};
+
 use ratatui::{
     backend::CrosstermBackend,
     layout::{Alignment, Constraint, Direction, Layout},
@@ -280,7 +280,7 @@ impl State {
             KeyCode::Char('s') if ctrl => {
                 self.switched_search_mode = true;
                 self.search_mode = self.search_mode.next(settings);
-                self.engine = crate::engines::engine(self.search_mode);
+                self.engine = engines::engine(self.search_mode);
             }
             KeyCode::Down if !settings.invert && self.results_state.selected() == 0 => {
                 return match settings.exit_mode {
@@ -780,7 +780,7 @@ pub async fn history(
             },
             context,
         },
-        engine: crate::engines::engine(search_mode),
+        engine: engines::engine(search_mode),
         results_len: 0,
         accept: false,
     };
