--- atuin/src/interactive/mod.rs
@@ -1,5 +1,0 @@
-pub mod cursor;
-pub mod search;
-
-mod history_list;
-mod inspector;
